PM Fitting now works (with c++17)make bin/WCSFit Also added verbose parameter for WCSFit to control output details.

+++ src/WCSFoF.cpp
@@ -426,14 +426,17 @@ main(int argc,
       vector<long> vid;
       extensionTable.readCell(key, "IDKEY", iextn);
       try {
+	/**/cerr << "Trying long read" << endl;
 	ft.readCells(vid, key);
       } catch (FTableError& m) {
 	// Trap for using int column in source file instead of long:
-	vector<int> vidint;
+	vector<LONGLONG> vidint;
+	/**/cerr << "Retrying int" << endl;
 	ft.readCells(vidint, key);
 	vid.reserve(vidint.size());
 	vid.insert(vid.begin(), vidint.begin(), vidint.end());
       }
+      /**/cerr << "Done" << endl;
       vector<bool> isStar(ft.nrows(), true);
       {
 	string starExpression;
